+++ carp_mobile_sensing/lib/sampling_packages/sensors/sensor_package.dart
@@ -16,10 +16,10 @@ class SensorSamplingPackage extends SmartphoneSamplingPackage {
   static const String NON_GRAVITATIONAL_ACCELERATION =
       CarpDataTypes.NON_GRAVITATIONAL_ACCELERATION_TYPE_NAME;
 
-  /// same as NON_GRAVITATIONAL_ACCELERATION, but averaged over a period
+  /// A set of acceleration (non-gravitational) features calculated over a
+  /// specific sampling period.
   ///  * Uses a [PeriodicSamplingConfiguration] for configuration.
-  static const String AVERAGE_NON_GRAVITATIONAL_ACCELERATION =
-      'averagenongravitationalacceleration';
+  static const String ACCELERATION_FEATURES = 'accelerationfeatures';
 
   /// Rotation of the device in x,y,z (typically measured by a gyroscope).
   ///  * Event-based measure.
@@ -61,13 +61,13 @@ class SensorSamplingPackage extends SmartphoneSamplingPackage {
             CarpDataTypes().types[CarpDataTypes.MAGNETIC_FIELD_TYPE_NAME]!),
         DataTypeSamplingScheme(
             DataTypeMetaData(
-              type: AVERAGE_NON_GRAVITATIONAL_ACCELERATION,
-              displayName: "Average Non Gravitational Accelerometer",
+              type: ACCELERATION_FEATURES,
+              displayName: "Accelerometer Features",
               timeType: DataTimeType.TIME_SPAN,
             ),
             PeriodicSamplingConfiguration(
               interval: const Duration(minutes: 1),
-              duration: const Duration(seconds: 2),
+              duration: const Duration(seconds: 3),
             )),
         DataTypeSamplingScheme(
             DataTypeMetaData(
@@ -88,8 +88,8 @@ class SensorSamplingPackage extends SmartphoneSamplingPackage {
         return AccelerometerProbe();
       case NON_GRAVITATIONAL_ACCELERATION:
         return UserAccelerometerProbe();
-      case AVERAGE_NON_GRAVITATIONAL_ACCELERATION:
-        return AverageUserAccelerometerProbe();
+      case ACCELERATION_FEATURES:
+        return AccelerometerFeaturesProbe();
       case MAGNETIC_FIELD:
         return MagnetometerProbe();
       case ROTATION:
@@ -105,6 +105,6 @@ class SensorSamplingPackage extends SmartphoneSamplingPackage {
 
   @override
   void onRegister() {
-    FromJsonFactory().register(AmbientLight());
+    FromJsonFactory().register(AmbientLight(0, 0, 0, 0));
   }
 }