+++ carp_mobile_sensing/lib/sampling_packages/sensors/sensors.dart
@@ -8,12 +8,15 @@ library sensors;
 
 import 'dart:async';
 import 'dart:io' show Platform;
+import 'dart:math' as math;
 
 import 'package:sensors_plus/sensors_plus.dart';
 import 'package:light/light.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:pedometer/pedometer.dart' as pedometer;
-import 'package:stats/stats.dart';
+// import 'package:stats/stats.dart';
+import 'package:statistics/statistics.dart';
+import 'package:sample_statistics/sample_statistics.dart';
 
 import 'package:carp_serializable/carp_serializable.dart';
 import 'package:carp_core/carp_core.dart';